# Conflicts:
#	public/js/SearchCharacter.js

+++ app/HTTP/Routing/CharacterController.php
@@ -48,7 +48,7 @@ class CharacterController extends AbstractController {
 
 			if (strtolower($extension) == 'png') {
 				$image = $post['Id'] . '.' . $extension;
-				unlink(__DIR__ . '/../../../public/images/characters/' . $image);
+				file_exists(__DIR__ . '/../../../public/images/characters/' . $image) && unlink(__DIR__ . '/../../../public/images/characters/' . $image);
 				$file->moveTo(__DIR__ . '/../../../public/images/characters/' . $image);
 			}
 		} else {
@@ -116,20 +116,23 @@ class CharacterController extends AbstractController {
 
 
 		try {
-			$this->characterManager->createCharacter($post);
+			$this->characterManager->createCharacter($post,$tags);
+			Flashes::add(FlashMessage::success("Le personnage {$post['Id']} a été créé !"));
 		} catch (\RuntimeException $e) {
 			Flashes::add(FlashMessage::danger($e->getMessage()));
 		}
 
+
+		/*
 		try {
 			$character = $this->characterManager->getByImage($post['Id']);
 			foreach ($tags as $tag) {
 				$this->characterTagDAO->create($character, $tag);
 			}
-			Flashes::add(FlashMessage::success("Le personnage {$post['Id']} a été créé !"));
+
 		} catch (\RuntimeException $e) {
 			Flashes::add(FlashMessage::danger($e->getMessage()));
-		}
+		}*/
 
 		return $res;
 	}
